refactor(spec): one unread-tensor refusal, and it names the loader that issued it
The refusal that keeps a newer drafter checkpoint from being served as an
older loader's architecture was private to the DFlash 1 loader. A second
drafter loader needs the same guard and the same wording, so it moves to the
shared speculative module and takes the loader's name as an argument: a
snapshot handed to the wrong generation's loader now says which one refused
it.

The test moves with it and gains an assertion on that name.
