+++ crates/rmlx-models/src/speculative/dflash/tests.rs
@@ -193,49 +193,3 @@ fn dflash_module_compiles() {
     let _snap = DFlashRoundState::snapshot;
     let _ = (_load, _bs, _walk, _snap);
 }
-
-// --- unread_tensor_refusal ---
-
-/// A snapshot every tensor of which was read loads; one carrying tensors this
-/// loader has no code for is refused, naming them.
-///
-/// Both directions, because the quiet direction is what keeps the check from
-/// decaying into noise a reader learns to skip. Mutation this fails on:
-/// `!consumed.contains(name)` -> `consumed.contains(name)`, which refuses the
-/// supported checkpoint and admits the unsupported one.
-#[test]
-#[allow(
-    clippy::unwrap_used,
-    reason = "test assertions: panicking on unexpected values is intentional"
-)]
-fn a_snapshot_this_loader_only_half_reads_is_refused_and_a_whole_one_is_not() {
-    let read_by_the_loader = ["fc.weight", "hidden_norm.weight", "norm.weight"];
-    let consumed: HashSet<String> = read_by_the_loader.iter().map(|s| (*s).to_owned()).collect();
-
-    // Every tensor consumed — the shape of the DFlash 1 checkpoint.
-    let whole: HashSet<String> = read_by_the_loader.iter().map(|s| (*s).to_owned()).collect();
-    assert!(
-        unread_tensor_refusal(&whole, &consumed).is_ok(),
-        "a snapshot the loader reads entirely must load"
-    );
-
-    // Extra weight families — the shape of the DFlash 2 checkpoint.
-    let mut partial = whole;
-    partial.insert("candidate_selector.successor_codebook".to_owned());
-    partial.insert("layers.0.attention_conv.base_kernel".to_owned());
-    let err = unread_tensor_refusal(&partial, &consumed).unwrap_err();
-    let msg = err.to_string();
-    assert!(
-        msg.contains('2'),
-        "refusal must count the unread tensors: {msg}"
-    );
-    assert!(
-        msg.contains("candidate_selector.successor_codebook")
-            && msg.contains("layers.0.attention_conv.base_kernel"),
-        "refusal must name the unread tensors: {msg}"
-    );
-    assert!(
-        !msg.contains("fc.weight"),
-        "a consumed tensor must not be reported unread: {msg}"
-    );
-}
